Move file functions to utils + refactor render priority area function
